requirements.txt
pandas==2.1.4
snowflake-connector-python==3.6.0
firebase-admin==6.4.0
python-dotenv==1.0.0
